Works with Raft
